<feat>(all): Sync from master

+++ src/main/java/com/webank/wecross/account/service/db/ChainAccountTableJPA.java
@@ -3,7 +3,7 @@ package com.webank.wecross.account.service.db;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ChainAccountTableJPA extends JpaRepository<ChainAccountTableBean, Long> {
+public interface ChainAccountTableJPA extends JpaRepository<ChainAccountTableBean, Integer> {
     List<ChainAccountTableBean> findByUsernameOrderByKeyIDDesc(String username);
 
     List<ChainAccountTableBean> findByIdentityOrderByKeyIDDesc(String identity);
